adding scaling font sizes

+++ app/models/chapter.rb
@@ -1,24 +1,39 @@
 class Chapter < ApplicationRecord
-  before_destroy :destroy_paths
+  before_destroy :getrid_children
 
-	has_many :plots
-	has_many :paths, through: :plots, dependent: :destroy
+	has_many :plots, dependent: :destroy
+	has_many :paths, through: :plots
 
  	accepts_nested_attributes_for :plots, :paths
     validates :content      , presence: true
     validates :name         , presence: true
 
-  after_destroy
+
    private
 
-   def destroy_paths
-   
-     self.paths.each do |t|
-     	t.destroy
-     end  
+   def getrid_children
+    chapterArray = []
 
-	
+    association_loop(self, chapterArray)
+    chapterArray.map{|r| r.destroy}
    end
+
+   def association_loop(g, chapterArray)
+
+    g.paths.each do |t|
+
+        if t.chapters.ids[0].present?
+          c = Chapter.find_by(pathPrev_id: t.id)     
+          chapterArray << t.chapters.first
+
+          if c    
+            association_loop(c, chapterArray)
+          end
+          t.destroy
+        end
+      end 
+     return chapterArray
+    end
 
 end
